Port EomApplication to GtkApplication
This removes the direct dependency on dbus and makes use of the
GApplication/GtkApplication facilities for uniqueness and activation.
Bump the glib requirement to 2.29.4, since we need
G_APPLICATION_NON_UNIQUE.

https://bugzilla.gnome.org/show_bug.cgi?id=622876

origin commit:
https://gitlab.gnome.org/GNOME/eog/commit/3d39587

Enable TotemScreensaver unconditionally

It uses GDBus instead of libdbus now. It can be enabled even if
libdbus is not available as GDBus/GIO is always present.

origin commit:
https://gitlab.gnome.org/GNOME/eog/commit/4c32882