Use new DBus API to show the current image in the file browser
This improves the functionality implemented with commit 9df5fd43.
The new API that will be included in Nautilus 3.3.4 not only opens a
view for the containing folder but also marks the image in the view.
The old behaviour is still available as fallback if the new API is
not offered on the system.

https://bugzilla.gnome.org/show_bug.cgi?id=650402

origin commit:
https://git.gnome.org/browse/eog/commit/?h=gnome-3-4&id=fa74473